backend/requirements.txt
fastapi
uvicorn
redis
python-socketio

httpx
pydantic-settings
python-jose
cryptography
passlib
bcrypt
python-multipart
websockets
